move scm storage print to instance_storage.go and add placeholder for control API Join
